requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
clr-loader==0.2.6
config==0.5.1
cryptography==41.0.3
cssselect==1.2.0
fire==0.5.0
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
metaphor-python==0.1.16
multidict==6.0.4
numpy==1.24.4
openai==0.28.0
opencv-python==4.8.0.76
pdf2docx==0.5.6
pdfminer.six==20221105
pdfquery==0.4.3
pycparser==2.21
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pypdf2==3.0.1
pyquery==2.0.0
python-docx==0.8.11
pythonnet==3.0.2
requests==2.31.0
roman==4.1
six==1.16.0
termcolor==2.3.0
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
